feat(backend): implement Pessoa CRUD operations

+++ backend/src/main/java/com/senai/prova/presentation/dtos/acao/CreatePessoaDTO.java
@@ -6,4 +6,4 @@ import jakarta.validation.constraints.NotNull;
 /**
  * @author MauricioMucci
  */
-public record CreatePessoaDTO(@NotNull Integer idPessoa, @NotBlank String mensagem) {}
+public record CreatePessoaDTO(@NotNull Long idPessoa, @NotBlank String mensagem) {}
